Refactor messenger client to have a cleaner api (#17)
## Why are we doing this?

The old interface was clunky. this better. And it's more consistent with how other node apps (like Express, a dependency) and the libraries mentioned in #15 look.

## Did you document your work?

I updated the documentation. And the interface is more in line with what Node people are used to.

## How can someone test these changes?

Steps to manually verify the change:

1. `nvm use; npm install`
2. `npm test` the test suite should pass
3. If you have the environment variables, `npm start` should start a working bot

## What possible risks or adverse affects are there?

* This is a pretty big refactor, but this is still pre-alpha. It's gradually shaping up though and now that I'm getting happier with the interface, it's in the "if you liked it you should have put a test on it" phase.

## What are the follow up tasks?

* add tests
* refactor. See "QUESTION: should we just send fat objects and use es6 destructuring to get what we want?" in the README

## Are there any known issues?

nothing new

## Did the test coverage decrease?

same stuff is covered. will increase as a followup.