Image echo (#18)
## Why are we doing this?

closes #5
This PR adds a `message.image` handler that just sends the same image back to the user.

## Did you document your work?

README and tests are updated.

## How can someone test these changes?

Steps to manually verify the change:

1. Run unit tests `npm test`
2. Run local bot
3. Talk to the bot and send a photo. Here's what my conversation looked like:

![screen shot 2016-10-21 at 3 17 02 pm](https://cloud.githubusercontent.com/assets/189908/19610384/e70a5330-97a1-11e6-9f3d-1f818070412c.png)


## What possible risks or adverse affects are there?

n/a

## What are the follow up tasks?

nothing directly related to this change.

## Did the test coverage decrease?

Since I documented some event changes, I added tests for them too. So coverage increased. But the old tests were accidentally running code they shouldn't have been hitting so the `send` method is no longer covered until it can be stubbed so it doesn't actually try and talk to Facebook.