Add sessions to the app (#38)
closes #33 

## Why are we doing this?

When the user messages us, we'll want to persist some state between messages AKA a session. The main goal is so we can associate a Look with a user across multiple messages.

## Did you document your work?

README is updated so events that send `session` are documented.

## How can someone test these changes?

Steps to manually verify the change:

In addition to unit tests...

1. `npm install`
2. start `npm run dev`
3. message the bot, if you repeat something (that isn't a command), the bot will say "stop repeating yourself" For example, you can repeatedly say "hi"

## What possible risks or adverse effects are there?

* ultimately, we'll need a persistent cache. That means there will be additional latency. That latency means we could run into race conditions where the bot essentially receives two messages at the same time. this shouldn't matter since our application is simple and doesn't need a lot of state, and the "last wins" approach is fine.

## What are the follow-up tasks?

* Version the cache keys
* I thought about wrapping the cache key generation in a function that's exportable, but I decided repeating myself was fine over introducing complexity

## Are there any known issues?

none (apart from the race condition already mentioned above)

## Did the test coverage decrease?

New code (except for POC) is tested, some new coverage added. Coverage goes from 63.2% to 67.48%