Add /ping so Departures knows the app is alive (#71)
## Why are we doing this?

To get this deployed (#6), we're using Departures. But it turns out that the Departures platform tests `/ping` or it kills the container. This adds that route with a 200 response so the deploy will satisfy the healthcheck.
On the [wiki](https://cnissues.atlassian.net/wiki/display/DEVOPS/3.+Deploy+with+Booking) click the link for "Troubleshooting Booking Deployments" which takes you do this [google doc](https://docs.google.com/document/d/1pv1VasQt641cmBTcLlkmqABXvo7HBTLOuIMbQUEMjzY/edit#)
On page 14 it says:

> Expose an HTTP healthcheck on /ping that returns a 2xx code. 

closes #71 

## Did you document your work?

Yeah, this kind of sticks out and is very Condé specific. So it'd be nice if it weren't in `app.js`, but that's the only place we define routes.

## How can someone test these changes?

Steps to manually verify the change:

1. `node index.js`
2. `curl -v localhost:3000/ping`
3. see `HTTP/1.1 200 OK`

## What possible risks or adverse effects are there?

There's slightly extra stuff in the code that's exists only to support Condé infrastructure.

## What are the follow-up tasks?

* We may want a better healthcheck, but that should live at another endpoint. Don't make healthchecks do double-duty.

## Are there any known issues?

none

## Did the test coverage decrease?

it's tested