Refactor with `node-config` (#86)
A step in the direction of streamlining our config setup. The user platform team is using node-config and this implementation replaces our custom config object with that module. It also moves the Messenger object to take the config as a constructor arg, reducing the calls to read and process the configuration values.

I also moved verifyRequestSignature into the Messenger class since that seemed appropriate.