Add Winston and log conversations to files and slack  (#98)
## Why are we doing this?

We have several loggers already, but none of them are good at the question, "what conversations are going on inside the bot?" This adds the [`winston`](https://github.com/winstonjs/winston) logger and adds it to our project #30. And as a test to justify Winston over a json file logger like Pino or Bunyan, adds a Slack output #60 to Winston. This also a followup to https://github.com/CondeNast/rkt-beauty-lenses/pull/97#issuecomment-266104574

> let's get more data... I'll make a spreadsheet with urls we've tried. This ties in with better logging so I'll continue with that for now.

This logging is also a fast way to grab what images are getting sent to the bot to debug if they're publicly accessible.

Without this logging, we'd either have to go through the Dashbot service to recreate conversations or create a custom view on top of the `debug` logs.

closes #30 
closes #60 

## Did you document your work?

New section of the README about logging.

## How can someone test these changes?

There are two feature toggles you have to enable (you can do this independently):
1. run `LOG_FILE=~/Desktop/beauty-lenses.log node index.js`
2. talk to your bot
3. there should be a `beauty-lenses.log` on your desktop with 1 json per file corresponding to the conversation
4. you can also do `tail -F ~/Desktop/beauty-lenses.log | jq .` and get some pretty json output as you talk to the bot

1. Do the same, but set `SLACK_WEBHOOK_URL` (the value is in `vault read secret/partnerships/accounts/slack`)
2. Behold, your chats should show in `#partnerships-zz-debug` in slack. When an unsupported image format comes in, the behavior is TBD. Ideally, a json blob gets spat out

## What possible risks or adverse effects are there?

If the logger blows up, it may mess up the main chat process. There may be a setting in https://github.com/winstonjs/winston#exceptions related to this.

## What are the follow-up tasks?

* Why are we sending `PAGE_ACCESS_TOKEN` to Dashbot?
* Move Dashbot config into `conversationLogger.js`?
* Improve logging output
* Save the timestamp we get from facebook?
* Our output doesn't have the `seq` incrementing int or the incoming message id. Having that would help debugging.

## Are there any known issues?

Not every message type is handled.
* [x] what happens if there's an error while sending to slack? Winston fails silently in the background.

## Did the test coverage decrease?

New code is covered.