Define a default look that is used in greeting responses (#130)
This is part of the quick reply based conversation flow desired by Product and described in #123. This PR adds support for specifying the default look via application configuration, add support for a welcome message with quick replies, and supports the quick replies with responses consistent with the existing looks and show me <look> commands.