Collapse together handling of postbacks, messages (text), and quick replies (#159)
This PR reduces code duplication and complexity. Text messages, postbacks, and quick replies are now all emitted at individual events and subsequently processed through the same response call. This allows us to have the processing divert later if needed but reduces code for the current state of the chat bot.