Create responses as types to hide boilerplate (#169)
## Why are we doing this?

I wanted to experiment with getting boilerplate out of the application code. I generally don't like thin wrappers, but I kept getting annoyed with the **readability** of trying to get the intent of the response from the raw object.

While I was rooting around, in accordance with discussion in #168, tried to make our code more consistently return a Promise.

## Did you document your work?

Some links to the relevant official docs added for response types.

## How can someone test these changes?

This is a refactor, everything should perform exactly the same as before.

## What possible risks or adverse effects are there?

Readability and code maintainability. Does this help or hurt?

## What are the follow-up tasks?

* Start treating more things as "types"
* add quickreplies?
* buttons elements?

## Are there any known issues?

none

## Did the test coverage decrease?

same