Hotfix work-around for quick reply processing (#179)
For a yet to be determined reason, quick reply payloads are coming through as mnqp_<FB_APP_ID>_<FB_PAGE_ID>_<PAYLOAD> instead of simply the payload that we are setting in the chatbot. This is causing the "we're not big on conversation" message to appear in situations where it should not because the processed text is falling through. See #176 for details. Since we do not know why this change occurred and cannot immediately pinpoint a code change that caused it, I am proposing this hotfix to have the chatbot working properly while we figure out the root cause.