Revert quick reply hot fix and add tests (#186)
This PR reverses the changes in #179. Per the discussion in #176, for a short period of time Facebook appeared to be sending quick reply payloads as a base 64 encoded string of data that included the Facebook application ID, page ID, and the intended payload. This change seems to have been reverted and the payload is being passed in plain text with no encoding.