Manually save session data & improve data structure (#185)
Implement a manual save session function and utilize it in applyLookToImage to save the data for operation. Additionally, this updates the structure of the saved data; instead of parallel arrays for the selfies, looks (final images) and registrations (image IDs), a single array of objects is kept. 
This also, as a side effect, provides a work around to #174, as no object is saved to the appliedLooks array when the image does not register or the look cannot be applied to the image.