Messenger SDK documentation (#189)
With the Beauty Lenses project and exploration of Facebook Messenger bots, one goal was to produce a reusable library that Partnerships/Accelerator or other Condé engineering teams could use as a springboard to build another Messenger bot. This PR adds documentation via a README for the SDK portion of the codebase to facilitate that goal.