Add a 'finish' event so tests take exactly as long as they need (#191)
## Why are we doing this?

Since the bot does things async, we're using a setTimeout in our tests to **guess** when Promises are finished. This changes things so we explicitly emit a `finish` event to signal we're done processing a request. But this diff only adds the event where we needed it for tests. This fixes the problem where tests were failing only some of the times. Tests now run exactly as long as they need to.

#### Prior art

It turns out sending an event is not an anti-pattern. With EventEmitter, sending events is all you got. It's common to send `error` events too (we currently just log them).

* `aiohttp` has [`write_eof`](http://aiohttp.readthedocs.io/en/stable/web_reference.html#aiohttp.web.StreamResponse.write_eof)
* `tornado` has [`finish`](http://www.tornadoweb.org/en/stable/web.html#tornado.web.RequestHandler.finish)
* Node's `http` has a [finish event](https://nodejs.org/api/http.html#http_event_finish)

closes #172 

![](https://qph.ec.quoracdn.net/main-qimg-a66ae0a1c1feab788a74783f01673aae-c?convert_to_webp=true)

## Did you document your work?

There's a blurb in the README that will cause a merge conflict with #189 

## How can someone test these changes?

Steps to manually verify the change:

Try repeatedly running tests. I think the easiest way is to:

1. `npm tdd`
2. cmd-s a file to cause the test watcher to run tests over and over
3. tests should pass all the time

## What possible risks or adverse effects are there?

The event is not applied consistently. It may be confusing to those reading the code why sometimes it emits `finish` and most of the time it does not. I hope the README blurb prevents that.

## What are the follow-up tasks?

* none

## Are there any known issues?

none, besides the impending merge conflict with #189 

## Did the test coverage decrease?

stays the same