Call getPublicProfile on greeting and store user profile in session (#202)
## Why are we doing this?

After we got some feedback from Facebook, we decided to reinstate greeting the user by name, which involves getting their public profile. Once we get their public profile, we also store it in the session with the key `profile`.

Closes #198

## Did you document your work?

Documentation is contained in the PDF report provided by Facebook; we decided via IRL conversation to re-add the personal greeting.

## How can someone test these changes?

* run the chat bot (`MODIFACE_API_URL=http://a0dev.modiface.com/ npm run dev`)
* say `hello`

![screen shot 2017-01-20 at 6 06 51 pm](https://cloud.githubusercontent.com/assets/1305168/22169635/3faf456e-df3b-11e6-8841-8af097b77850.png)

## What possible risks or adverse effects are there?

It is possible that the call to get the public profile could fail (returning `{}`) or that the user has no valid first name (blank? spaces?). These basic cases are handled, but there are perhaps others that might trip up the chat bot. I have not tested an exhaustive variety of first name possibilities.

## What are the follow-up tasks?

None.

## Are there any known issues?

None.

## Did the test coverage decrease?

No, added tests to cover the base cases.