Add replies to stickers for more personality (#209)
## Why are we doing this?

Facebook suggested that the bot should reply to stickers. Adding a response to stickers gives the bot a little bit of personality. For now this PR adds naive emoji reactions to any stickers received.

closes #196 

## Did you document your work?

Not necessary for this feature.

## How can someone test these changes?

Steps to manually verify the change:

1. Send a "thumbsup" to the bot
2. Send a sticker
3. the bot should send back a response when you do this

screenshot:
<img width="380" alt="screen shot 2017-01-23 at 3 49 55 pm" src="https://cloud.githubusercontent.com/assets/189908/22224471/67016700-e184-11e6-89b9-bea833714f98.png">

## What possible risks or adverse effects are there?

* increases non-looks related noise in logs

## What are the follow-up tasks?

* copy edits to what the responses should be. I used https://emojikeyboard.org/ to pick some arbitrary responses.

## Are there any known issues?

Sentiment is not taken into account. The emoji response is the same no matter if it receives a happy or a sad sticker.

## Did the test coverage decrease?

stays the same. new paths are covered.