Random cleanup of SDK code (#211)
## Why are we doing this?

This is some minor cleanup to polish the Messenger SDK portion of this code. There are some deletions of dead code, and some tweaks. For SDK code, I wanted to decouple it from `config`, so that's why dashbot config changes. This is one of the rare cases where you should really just look at the diff.

## Did you document your work?

Env config docs now moved to the SDK README.

## How can someone test these changes?

Steps to manually verify the change:

1. This is a refactor, nothing should change

## What possible risks or adverse effects are there?

None

## What are the follow-up tasks?

None

## Are there any known issues?

None

## Did the test coverage decrease?

Sort of increases because of the deletions.