Add basic Google Sheet integration (#5)
## Why are we doing this?

For ease of input, we're using Google Spreadsheets as our source for data on these quizzes. This PR adds the plumbing for getting data from all the wookbooks in a Spreadsheet into the bot. And provides a mechanism for updating the data without a restart.

Also does some cleanup that's natural in this early phase.

closes #4 

## Did you document your work?

The GOOGLE env vars take some imagination. To get a fresh set of keys, I made a page on our wiki: https://cnissues.atlassian.net/wiki/display/PA/Google+Developers

## How can someone test these changes?

Steps to manually verify the change:

1. `npm install`
2. Run the bot `npm run dev`
3. message the bot with "fetch quizzes", if you have google creds, you'll get  a success message. If you don't, you'll get a basic error message, but the console should have useful information why it failed.

<img width="280" alt="screen shot 2017-01-29 at 8 36 45 pm" src="https://cloud.githubusercontent.com/assets/189908/22410831/bb5fa400-e662-11e6-82e1-ed45736130cc.png">


## What possible risks or adverse effects are there?

* none

## What are the follow-up tasks?

None

## Are there any known issues?

* Anyone can trigger a refresh
* local version of quiz is local. A load balanced version of the bot should be fine and still possible (sessions would need another backend too)

## Did the test coverage decrease?

Coverage isn't great, but the google-spreadsheet package is hard to test around. Future functionality should be testable though with the hard stuff in it's own place.