Add a proof of concept quiz (#13)
## Why are we doing this?

This is the first iteration of a quiz. Users can answer questions, and they get a grade when they're done.

#### TODO
- [x] shuffle new element into a `customObjects` or `customResponses` or `customElements`
- [x] shuffle some lines to make the diff smaller
* [x] document the application session object

## Did you document your work?

The conversation flow is not documented because it's still really early. Eventually, it should be natural enough to not need documenting, and the docs that will exist will be some walkthroughs on a wiki.

## How can someone test these changes?

Steps to manually verify the change:

1. start a bot
2. say "hi"
3. follow the quiz
4. You're now Dave Coulier
5. (optional) delete a photo and generate 2-ups again with `node scripts/two-up.js`

## What possible risks or adverse effects are there?

None

## What are the follow-up tasks?

None

## Are there any known issues?

Don't expect it to be smart or fancy. This is just the first one.

## Did the test coverage decrease?

Conversation flow is not tested since we don't know what behavior we quite want yet, but library code is tested.