Optionally check for & emit greetings in the SDK (#18)
Closes #17 

## Why are we doing this?

The majority of _best practices_ for bots indicate that a bot should always respond to a greeting in some form. With _Beauty Lenses_, we implemented a regular expression check against common greetings in the application layer to handle this.

This PR moves some of the greeting logic to the SDK and introduces a `message.greeting` event. Application code surrounding greetings becomes simpler. Emitting greetings as independent events, as opposed to `message.text` events in prior versions, is done optionally, with the default turned on. This encourages application developers to support responses to the event, without forcing existing code, in _Beauty Lenses_ for example, to drastically change if they have already supported greetings in their `message.text` processing.
 
*TODO*:
* ~~tests in `main.spec.js`~~ `test/messenger/app.spec.js` covers this for now, and adding them at the application level would require a lot of quiz loading workaround that we can explore in the near future.

## Did you document your work?

The SDK `README` has been updated with details on the `message.greeting` event.

## How can someone test these changes?

* check out the `sdk-greetings/#17` branch
* run the chat bot
* say `hello` or `hi`
* observe the response and start a quiz!

## What possible risks or adverse effects are there?

The greetings regex is certainly not comprehensive, and there is no means for a user to add to the expression, currently. The alternative is to disabled `emitGreetings` in the `Messenger` object and rely on the application to implement a handler, as _Beauty Lenses_ does.

## What are the follow-up tasks?

* How does this overlap w/#15? Is there a common pattern that these two issues should converge upon?
* #26 and #28, from ideas in comments on this PR

## Are there any known issues?

None.

## Did the test coverage decrease?

The new event is covered ~~could use better coverage on the branch when `emitMessage` is false~~.

Before:
```
=============================== Coverage summary ===============================
Statements   : 72.15% ( 171/237 )
Branches     : 69.89% ( 65/93 )
Functions    : 77.27% ( 17/22 )
Lines        : 73.36% ( 168/229 )
================================================================================
```

After:
```
=============================== Coverage summary ===============================
Statements   : 72.61% ( 175/241 )
Branches     : 70.3% ( 71/101 )
Functions    : 77.27% ( 17/22 )
Lines        : 73.82% ( 172/233 )
================================================================================
```