Correcting some 'message.greeting' events to 'text.greeting' (#34)
## Why are we doing this?

These changes were earmarked for #18, but got lost in a giccup.

## Did you document your work?

Documentation is present in the previous PR, #18.

## How can someone test these changes?

I would say `npm test`, but I am finding that the event tests run afoul in some cases.

For example, if I change the `message.text` events in `./src/app.spec.js` to `text` in the greeting test (not in this diff), I got what appear to be some conflicts amongst the events that lead to failed tests.

Try it:
```
index ea29453..0837bcd 100644
--- a/test/messenger/app.spec.js
+++ b/test/messenger/app.spec.js
@@ -221,8 +221,8 @@ describe('app', () => {
         assert.equal(payload.senderId, 'senderId');
       });
 
-      messenger.once('message.text', (payload) => {
-        assert.fail('message.text', 'text.greeting', 'incorrect event emitted');
+      messenger.once('text', (payload) => {
+        assert.fail('text', 'text.greeting', 'incorrect event emitted');
       });
 
       messenger.onMessage(event, {});
```

## What possible risks or adverse effects are there?

Other than the concerns about a test w/modifications that are not in this PR, this should not have any adverse effects of risks.

## What are the follow-up tasks?

I want to dig into the tests to see what's up. This is the second time this kind of thing has come up, it surfaced on _Beauty Lenses_ also.

## Are there any known issues?

Nothing that hasn't already been mentioned.

## Did the test coverage decrease?

Stays the same.