Emit cleaned (whitespace) names with `text.greeting` event (#37)
Closes #26 

## Why are we doing this?

This provides a similar interface to Facebook and improves application developer friendliness specific to the `text.greeting` event. When responding to the event, a developer can now reference the (existing) public profile information inside the session (`session.profile.first_name` for example) or use a whitespace cleaned version directly with `firstName`. In addition to `firstName`, `surName`, and `fullName` are provided; `fullName` is the `surName` appended to the `firstName` with a single, separating space.

## Did you document your work?

Updated the `text.greeting` event detail in the SDK `README`

## How can someone test these changes?

* run the chat bot, `npm run dev`
* send it a greeting, such as `hi` or `hello`
* observe the sample response that makes use of the `firstName` and `fullName` attributes

## What possible risks or adverse effects are there?

Data duplication: we now have the names in two places for this event, the `session` and at the top level. We are choosing developer convenience over a little bit of duplication.

## What are the follow-up tasks?

None.

## Are there any known issues?

None identified.

## Did the test coverage decrease?

No, the few new lines are covered. Opted not to test the functionality of `.trim()`