Bots should respond to help requests / `text.help` in SDK (#35)
Closes #33, minus any copy updates that come later.

## Why are we doing this?

Responding to a request for help is a basic function that a bot should respond to with useful information. This PR implements a `text.help` event, similar to `text.greeting`. In addition, this implements the suggestion in #28 with regard to help; allowing the application to override the regular expression via the optional parameter.

## Did you document your work?

Yes, the SDK `README` contains updates covering the `text.help` event and the `emitHelp` option.

## How can someone test these changes?

* run the bot: `npm run dev`
* sent a message starting with _help_, observe the help response from the messages 
* update the application such that the value for `emitHelp` contains a different, valid regular expression
* run the bot again and send a message that will match the new regular expression
* optionally, set `emitHelp` to false and see that the help response is not sent

## What possible risks or adverse effects are there?

Adding another event adds complexity to the SDK. It should reduce complexity in the application and encourage application developers to implement a useful help response. The default regular expression will miss things like, "_I need help!_", but that purpose can be served by overriding it. We can also monitor messages via DashBot and update the regular expression accordingly.

## What are the follow-up tasks?

If we like and merge the implementation of `emitHelp` as a regular expression, extend it to `emitGreeting` to satisfy #28 

## Are there any known issues?

None identified.

## Did the test coverage decrease?

Coverage is slightly increased as the new code is covered.