Basic namespacing of cache keys using the application id (#39)
Closes #2 

## Why are we doing this?

As we explore multiple bot projects, we have discussed the ability to deploy multiple bots to a single page or a mechanism for distributed caching across multiple bots, regardless of whether or not they are on deployed on the same page. Both of these potential developments require a way to differentiate a user's cache between multiple applications. Currently, the cache keys are simply the Facebook Messenger ID, and this PR adds a simple `getCacheKey` function to the SDK which prepends provided data with the application ID. For `rkt-pop-quiz-hotshot`, cache keys will now be application ID, '-', sender ID.

## Did you document your work?

~~Yes, the SDK README contains additional details about the session key scheme and `getCacheKey` function.~~ Purposefully skipping since this is purely internal to the SDK.

## How can someone test these changes?

Running the bot normally and observing that session states are get and set correctly. Logging the `quizState` might be the best check for this in the current application.

## What possible risks or adverse effects are there?

If we were already running a scheme with a distributed cache, the new scheme would invalidate old sessions. This also makes it very easy to change the existing scheme again, so the risk of session invalidation persists. That said, this is a low risk given that we are not using anything distributed for session storage and are okay with losing application sessions when we redeploy.

## What are the follow-up tasks?

* discussion on the optimal cache scheme to meet long-term needs and any code updates that are required

## Are there any known issues?

None.

## Did the test coverage decrease?

No.