Move config into a local singleton (#7)
## Why are we doing this?

I'm trying to extract all of the shared state into something that can be accessed across modules. Right now the `this.config` inside `Messenger` is the majority of the shared state. In talking this through with @stripethree, we realized there might be an issue with non-namespaced configuration as well. This introduces `launch-vehicle-fbm.` as the prefix for all configuration variables.

## Did you document your work?

Haven't yet – there currently isn't any configuration in the `README.md` so nothing's changed.

## How can someone test these changes?

1. `npm i`
2. `npm t`
3. No test failures and three new tests under the `config` module

## What possible risks or adverse effects are there?

None, assuming the test coverage has caught edge cases for configuration changes. A few tests failed with the change, but that was caused by the constructor changing.