Change so the app, not sdk, supplies translation messages (#19)
## Why are we doing this?

The text translation layer is supposed to be maintained by the application, but it's currently locked into the SDK. This adds a new dependency, [`app-root-dir`](https://www.npmjs.com/package/app-root-dir) to move `messages.js` from relative to the SDK to relative to the user's app.

`app-root-dir` is fairly simple, but I'd rather just add a new dependency than rewrite it ourselves. It has no dependencies, and it looks like the packages that depend on it are solving the same problem we're trying to solve. For example, [lasso-i18n](https://github.com/lasso-js/lasso-i18n/blob/master/index.js) uses it to pull a template.

closes #14

## Did you document your work?

The README is updated, and the TBD section is now gone.

## How can someone test these changes?

Steps to manually verify the change:

There's just unit tests for now, but you can `npm link` in this repo and edit https://github.com/CondeNast/rkt-pop-quiz-hotshot/pull/47 , follow the new README here, and try it out

## What possible risks or adverse effects are there?

* There is a `require` that goes outside the sdk's control. This is essentially trusting user content, but it's the user's own code.

## What are the follow-up tasks?

* none

## Are there any known issues?

none

## Did the test coverage decrease?

increases, the loading code wasn't covered before