Conversion of conversation logger to a standalone module (#23)
## Why are we doing this?

`ConversationLoggger` is a viable, stand-alone, (open source)? module and this does some cleanup that will help maintainability and open to the door to it being separated from the SDK should we decide that in the future.

## Did you document your work?

No additional documentation added. There is no standalone `README` for this and these changes are all internal to the SDK so nothing that is currently documented in the `README` changes. The minimal changes to the tests should back this up.

## How can someone test these changes?

`npm i`
`npm t`

Testing against a skeleton bot:
* link this library via `npm link`
* use [this gist] as a test bot
* run the bot!

[this gist]: https://gist.github.com/stripethree/3fa81ea549957502bd0c1b7b85ae5abf

## What possible risks or adverse effects are there?

While this PR touts decoupling this as a standalone module, the configuration namespacing is still tied to the SDK. Thus, it also adds additional configuration elements; these do not have to be present and `has` is used in the `ConversationLogger` constructor to prevent exceptions from bubbling up. `Messenger` does not access any of the `ConversationLogger` configuration items.

## What are the follow-up tasks?

* Do we want to decouple this from Messenger entirely?
* Do we want to have this be event driven?

## Are there any known issues?

None identified; ~~integration testing pending~~.

## Did the test coverage decrease?

No.