Fix how this.options was referencing wrong object in conversationLogger (#28)
## Why are we doing this?

In the conversation logger refactor #23 , a global options became an instance attribute `this.options`. Unfortunately, because Winston calls `this.slackFormatter` and not the `ConversationLogger` instance, `this.options` changed unexpectedly to reference the Winston logger and was `undefined` when we meant for it to reference `ConversationLogger.options`.

## Did you document your work?

Some additional Flow type docs added while trying to debug this.

## How can someone test these changes?

Steps to manually verify the change:

1. Run the bot in a dev environment (`npm link`)
2. You don't need to enable Slack config, but it's nice
3. Messaging the bot should not result in a `TypeError: Cannot read property 'slackChannel' of undefined`

## What possible risks or adverse effects are there?

* I may have missed another `this` for something else, but we can tackle that when it shows

## What are the follow-up tasks?

* none

## Are there any known issues?

none

## Did the test coverage decrease?

Increases with Flow type. Unfortunately, the integration with Winston is still uncovered because we don't have integration tests for this, just unit tests.