Make response factories easier to access (#31)
## Why are we doing this?

Currently, if users want to use the response object factories, they have to reach into the source (which also currently requires knowledge of the source tree). This adds documentation for how to use them and also makes them easier to access.

closes #21 

## Did you document your work?

Some major rework of the README

## How can someone test these changes?

Steps to manually verify the change:

Just unit tests (`main.spec.js`). You can npm link this in to a side project but that's going above what you should need to do to test this.

## What possible risks or adverse effects are there?

* none

## What are the follow-up tasks?

* I noted that quick replies are still a WIP, so we'll have to revisit that one day

## Are there any known issues?

* If the user assumes they can tree shake `require('launch-vehicle-fbm/responses/Text')`, they'll be sad. But this isn't documented so that's ok.

## Did the test coverage decrease?

new code is covered