Remove example key-value pairs that do not apply to the SDK (#35)
## Why are we doing this?

We copy-pasted the `example.env` file from a bot project using the SDK and through review & release, we missed that there are some extraneous values present. These values are not relevant to the SDK, so they are being removed to avoid any confusion.

## Did you document your work?

No additional documentation required; the `README` covers the applicable values that are not impacted by this PR and does not mention any of the removed values

## How can someone test these changes?

`npm i`
`npm t` - the tests use `example.env`

## What possible risks or adverse effects are there?

None identified.

## What are the follow-up tasks?

None identified.

## Are there any known issues?

None identified.

## Did the test coverage decrease?

No.